v. 0.1.0

+++ Pod/Classes/DSOHashery.m
@@ -212,7 +212,7 @@
         result += n * pow(_base, power);
     }
     
-    NSLog(@"original base-10 number is: %ld", result);
+    NSLog(@"original base-10 number is: %lg", (long)result);
     return result;
 }
